requirements.txt
fastapi==0.111.0
pillow==10.3.0
qdrant-client==1.9.1
requests==2.32.3
torch==2.3.0+cu121
transformers==4.41.2
uvicorn==0.30.0
